refactor: pull DevMockClient out of src/lib/ai/clientFactory
src/lib/ai/clientFactory.ts statically imported mockStateManager from
src/lib/devtools/ and MockScenarios from src/lib/ai/__mocks__/, dragging
test fixtures and devtools into every production AI consumer's module
graph. Move the DevMockClient class to src/lib/devtools/devMockClient.ts
and expose a tryCreateDevMockClient() factory. clientFactory now lazy
requires it via the existing dev-only branch, so production bundles that
never hit that branch don't pull devtools/__mocks__ into the AI graph.

Verified all clientFactory tests (browser + mockIntegration) still pass —
jest.mock() targets the underlying module path, which both clientFactory
(via require) and devMockClient (via static import) resolve to the same
file.
